+++ crates/rolldown_plugin_vite_dynamic_import_vars/src/ast_visit_2.rs
@@ -158,7 +158,7 @@ impl<'ast> DynamicImportVarsVisit<'ast, '_> {
       self
         .magic_string
         .get_or_insert_with(|| MagicString::new(self.source_text))
-        .update(import_expr.span.start as usize, import_expr.span.end as usize, replacement)
+        .update(import_expr.span.start, import_expr.span.end, replacement)
         .expect("update should not fail in dynamic import vars plugin");
 
       self.need_helper = true;
