+++ crates/rolldown_plugin_vite_import_glob/src/utils_2.rs
@@ -256,7 +256,7 @@ impl<'ast> GlobImportVisit<'_> {
     self
       .magic_string
       .get_or_insert_with(|| string_wizard::MagicString::new(self.code))
-      .update(span.start as usize, span.end as usize, replacement)
+      .update(span.start, span.end, replacement)
       .expect("update should not fail in import glob plugin");
   }
 }
